iScript: moved anti-afk functionality into iScript. Anti-afk will only occur while a plugin implementing iScript framework is running. Once plugin is stopped anti-afk will also stop. This is to prevent scenarios where a plugin may stop or crash and player would AFK indefinitely until 6 hour logout or user intervened.
iutils: anti-afk functionality removed from iUtils (see above)
