+++ iutils/src/main/java/net/runelite/client/plugins/iutils/scripts/iScript.java
@@ -1,10 +1,16 @@
 package net.runelite.client.plugins.iutils.scripts;
 
 import lombok.extern.slf4j.Slf4j;
+import net.runelite.api.events.GameTick;
+import net.runelite.client.eventbus.Subscribe;
 
+import java.util.Random;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
+
+import static java.awt.event.KeyEvent.VK_LEFT;
+import static java.awt.event.KeyEvent.VK_RIGHT;
 
 @Slf4j
 public abstract class iScript extends UtilsScript {
@@ -68,6 +74,37 @@ public abstract class iScript extends UtilsScript {
     @Override
     protected void shutDown() {
         stop();
+    }
+
+    private void checkIdleLogout() {
+        // Check clientside AFK first, because this is required for the server to disconnect you for being afk
+        int idleClientTicks = game.client.getKeyboardIdleTicks();
+
+        if (game.client.getMouseIdleTicks() < idleClientTicks) {
+            idleClientTicks = game.client.getMouseIdleTicks();
+        }
+
+        if (idleClientTicks > 12500) {
+            Random r = new Random();
+            log.info("iScript Resetting idle");
+
+            if (r.nextBoolean()) {
+                game.pressKey(VK_LEFT);
+            } else {
+                game.pressKey(VK_RIGHT);
+            }
+
+            game.client.setKeyboardIdleTicks(0);
+            game.client.setMouseIdleTicks(0);
+        }
+    }
+
+    @Subscribe
+    private void onGameTick(GameTick event) {
+        if (!started()) {
+            return;
+        }
+        checkIdleLogout();
     }
 
     protected abstract void loop();
